+++ power_iteration.h
@@ -20,15 +20,15 @@
 #include "leading_eigenpair.h"
 #include "modularity_matrix.h"
 
-void power_iteration(modularity_matrix* mod_matrix, double* eigen_vector);
-double* generate_rand_vec(double* vec_0, int vec_size);
-void get_next_vec(modularity_matrix* mod_matrix, double* vec, double* next_vec, int vec_size);
-int smaller_than_eps(double* vec, double* next_vec, int vec_size);
-double calc_multiplication(modularity_matrix* mod_matrix, double *vec, int row, bool to_shift);
-double mult_sparse_mat_row_by_vec(modularity_matrix* mod_matrix, double *vec, int row);
-double dot_product(double* vector_1, double* vector_2, int vec_size);
-double int_dot_product(int* vector_1, double* vector_2, int vec_size);
-double sum_vec(double* vector, int vec_size);
+void power_iteration(modularity_matrix* mod_matrix, float* eigen_vector);
+float* generate_rand_vec(float* vec_0, unsigned int vec_size);
+void get_next_vec(modularity_matrix* mod_matrix, float* vec, float* next_vec, unsigned int vec_size);
+unsigned int smaller_than_eps(float* vec, float* next_vec, unsigned int vec_size);
+float calc_multiplication(modularity_matrix* mod_matrix, float *vec, unsigned int row, bool to_shift);
+float mult_sparse_mat_row_by_vec(modularity_matrix* mod_matrix, float *vec, unsigned int row);
+float dot_product(float* vector_1, float* vector_2, unsigned int vec_size);
+float int_dot_product(unsigned int* vector_1, float* vector_2, unsigned int vec_size);
+float sum_vec(float* vector, unsigned int vec_size);
 
 
 #endif /* POWER_ITERATION_H_ */
